Updated autotests for NetworkInterface
