+++ test/commands/arm/network/arm.network.nic-tests.js
@@ -1,4 +1,4 @@
-﻿/**
+/**
  * Copyright (c) Microsoft.  All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,166 +13,183 @@
  * limitations under the License.
  */
 
-// Warning: This code was generated by a tool.
-// 
-// Changes to this file may cause incorrect behavior and will be lost if the
-// code is regenerated.
-
 'use strict';
 
 var should = require('should');
 var util = require('util');
 var _ = require('underscore');
+var fs = require('fs');
 
-var testUtils = require('../../../util/util');
 var CLITest = require('../../../framework/arm-cli-test');
 var utils = require('../../../../lib/util/utils');
-var NetworkTestUtil = require('../../../util/networkTestUtil');
 var tagUtils = require('../../../../lib/commands/arm/tag/tagUtils');
-var networkUtil = new NetworkTestUtil();
+var testUtils = require('../../../util/util');
+
+var networkTestUtil = new (require('../../../util/networkTestUtil'))();
+var vmTestUtil = new (require('../../../util/vmTestUtil'))();
 
 var generatorUtils = require('../../../../lib/util/generatorUtils');
 var profile = require('../../../../lib/util/profile');
 var $ = utils.getLocaleString;
 
 var testPrefix = 'arm-network-nic-tests',
-  groupName = 'xplat-test-nic',
-  location;
-var index = 0;
+  groupName = 'xplat-test-nic-custom',
+  location,
+  poolName = 'test-pool',
+  inboundRuleName = 'test-inbound-rule',
+  protocol = 'tcp',
+  frontendPort = '90',
+  backendPort = '90',
+  vmssParams;
 
 var networkInterfaces = {
-  privateIPAddress: '10.0.0.42',
+  group: groupName,
+  name: 'test-nic',
+  privateIPAddress: '10.0.0.22',
+  newPrivateIPAddress: '10.0.0.25',
   privateIPAddressVersion: 'IPv4',
-  ipConfigurationName: 'createipconf',
+  newPrivateIPAddressVersion: 'IPv6',
   internalDnsNameLabel: 'internal-dns-foo',
-  internalDnsNameLabelNew: 'internal-dns-bar',
-  enableAcceleratedNetworking: 'false',
-  enableAcceleratedNetworkingNew: 'true',
-  enableIPForwarding: 'false',
-  enableIPForwardingNew: 'true',
-  location: 'westus',
-  name: 'networkInterfaceName'
+  newInternalDnsNameLabel: 'internal-dns-bar',
+  ipConfigName: 'another-ip-config',
+  enableIpForwarding: false,
+  newEnableIpForwarding: true,
+  tags: networkTestUtil.tags,
+  newTags: networkTestUtil.newTags,
+  attachedVMName: 'tempXplatVMForNicTests',
+  attachedVMStorageAccount: 'xplateffectiveaccount',
+
+  location: 'southeastasia',
+  virtualmachineIndex: '0',
+  networkInterfaceName: 'test',
+  ipConfigurationName: 'test',
+  vmssName: 'nicVmssName'
 };
 
-networkInterfaces.virtualNetworkName = 'virtualNetworkName';
-networkInterfaces.subnetName = 'subnetName';
-networkInterfaces.publicIPAddressName = 'publicIPAddressName';
+networkInterfaces.virtualNetworkName = 'test-vnet';
+networkInterfaces.subnetName = 'test-subnet';
 networkInterfaces.networkSecurityGroupName = 'networkSecurityGroupName';
-networkInterfaces.networkInterfaceName = 'networkInterfaceName';
 
-var subnet = {
-  addressPrefix: '10.0.0.0/16',
-  addressPrefixNew: '10.0.0.0/24'
-};
 var virtualNetwork = {
-  location: 'westus'
+  name: 'test-vnet',
+  location: 'southeastasia'
+};
+var subnet = {
+  name: 'test-subnet',
+  addressPrefix: '10.0.0.0/16',
+  virtualNetworkName: 'test-vnet'
 };
 var networkSecurityGroup = {
-  location: 'westus'
+  name: 'networkSecurityGroupName',
+  location: 'southeastasia'
 };
-var publicIPAddress = {
-  location: 'westus'
+var vmTestUtilCreateVMSSWithParamFile = {
+  vmssName: 'nicVmssName',
+  paramFile: 'test/data/vmssParamTestNIC.json',
+  storageAccount: 'xplatteststorage1',
+  storageCont: 'xplatteststoragecnt1',
+  publicIpName: 'xplattestip',
+  domainNameLabel: 'vmssiplabel',
+  idleTimeout: '30',
+  nicDnsServer: '10.11.12.13'
 };
-var invalidPrivateIPAddress = {
-  privateIPAddress: '10.11.12.13.14',
-  location: 'westus',
-  virtualNetworkName: 'virtualNetworkName',
-  subnetName: 'subnetName',
-  publicIPAddressName: 'publicIPAddressName',
-  name: 'InvalidPrivateIPAddressName',
-  group: groupName
+
+var ipConfigProp1 = {
+  name: 'config01'
 };
-var ipVersionOutOfRange = {
-  privateIPAddressVersion: 'IP',
-  location: 'westus',
-  virtualNetworkName: 'virtualNetworkName',
-  subnetName: 'subnetName',
-  publicIPAddressName: 'publicIPAddressName',
-  name: 'IPVersionOutOfRangeName',
-  group: groupName
+var ipConfigProp2 = {
+  name: 'config02',
+  privateIPAddress: '10.0.0.26',
+  isPrimary: true
 };
-var invalidIPForwarding = {
-  enableIPForwarding: 'Disable',
-  location: 'westus',
-  virtualNetworkName: 'virtualNetworkName',
-  subnetName: 'subnetName',
-  publicIPAddressName: 'publicIPAddressName',
-  name: 'InvalidIPForwardingName',
-  group: groupName
+var publicIpProp = {
+  name: 'test-ip',
+  domainName: 'foo-domain-test',
+  allocationMethod: 'Static',
+  ipVersion: 'IPv4',
+  idleTimeout: 4,
+  tags: networkTestUtil.tags
 };
-var attachNsgToNic = {
-  location: 'westus',
-  virtualNetworkName: 'virtualNetworkName',
-  subnetName: 'subnetName',
-  publicIPAddressName: 'publicIPAddressName',
-  networkSecurityGroupName: 'networkSecurityGroupName',
-  name: 'AttachNsgToNicName',
-  group: groupName
-};
-var detachNsgFromNic = {
-  location: 'westus',
-  networkSecurityGroupName: 'networkSecurityGroupName',
-  name: 'AttachNsgToNicName',
-  group: groupName
+var lbProp = {
+  name: 'test-lb',
+  fipName: 'test-fip',
+  publicIpProp: {
+    name: 'test-fip-ip',
+    domainName: 'baz-domain-test',
+    allocationMethod: 'Dynamic',
+    ipVersion: 'IPv4',
+    idleTimeout: 4,
+    tags: networkTestUtil.tags
+  }
 };
 
 var requiredEnvironment = [{
   name: 'AZURE_VM_TEST_LOCATION',
-  defaultValue: 'westus'
+  defaultValue: 'southeastasia'
 }];
 
 describe('arm', function () {
   describe('network', function () {
     var suite, retry = 5;
     var hour = 60 * 60000;
+    var testTimeout = hour;
 
     before(function (done) {
-      this.timeout(hour);
+      this.timeout(testTimeout);
       suite = new CLITest(this, testPrefix, requiredEnvironment);
       suite.setupSuite(function () {
         location = networkInterfaces.location || process.env.AZURE_VM_TEST_LOCATION;
         groupName = suite.isMocked ? groupName : suite.generateId(groupName, null);
+
         networkInterfaces.location = location;
         networkInterfaces.group = groupName;
         networkInterfaces.name = suite.isMocked ? networkInterfaces.name : suite.generateId(networkInterfaces.name, null);
+
+        ipConfigProp1.group = groupName;
+        ipConfigProp1.nicName = networkInterfaces.name;
+
+        ipConfigProp2.group = groupName;
+        ipConfigProp2.nicName = networkInterfaces.name;
+        ipConfigProp2.vnetName = networkInterfaces.virtualNetworkName;
+        ipConfigProp2.subnetName = networkInterfaces.subnetName;
+
+        publicIpProp.location = location;
+        publicIpProp.group = groupName;
+        publicIpProp.name = suite.isMocked ? publicIpProp.name : suite.generateId(publicIpProp.name, null);
+
+        lbProp.location = location;
+        lbProp.group = groupName;
+        lbProp.name = suite.isMocked ? lbProp.name : suite.generateId(lbProp.name, null);
+        lbProp.publicIpProp.location = location;
+        lbProp.publicIpProp.group = groupName;
+
+        poolName = suite.isMocked ? poolName : suite.generateId(poolName, null);
+        inboundRuleName = suite.isMocked ? inboundRuleName : suite.generateId(inboundRuleName, null);
+
+        vmssParams = JSON.parse(fs.readFileSync(vmTestUtilCreateVMSSWithParamFile.paramFile));
+
         if (!suite.isPlayback()) {
-          networkUtil.createGroup(groupName, location, suite, function () {
-            var cmd = 'network vnet create -g {1} -n virtualNetworkName --location {location} --json'.formatArgs(virtualNetwork, groupName);
+          networkTestUtil.createGroup(groupName, location, suite, function () {
+            var cmd = 'network vnet create -g {1} -n {name} --location {location} --json'.formatArgs(virtualNetwork, groupName);
             testUtils.executeCommand(suite, retry, cmd, function (result) {
               result.exitStatus.should.equal(0);
               var output = JSON.parse(result.text);
-              invalidPrivateIPAddress.virtualNetworkId = suite.isMocked ? output.id : suite.generateId(invalidPrivateIPAddress.virtualNetworkId, null);
-              ipVersionOutOfRange.virtualNetworkId = suite.isMocked ? output.id : suite.generateId(ipVersionOutOfRange.virtualNetworkId, null);
-              invalidIPForwarding.virtualNetworkId = suite.isMocked ? output.id : suite.generateId(invalidIPForwarding.virtualNetworkId, null);
-              attachNsgToNic.virtualNetworkId = suite.isMocked ? output.id : suite.generateId(attachNsgToNic.virtualNetworkId, null);
-              detachNsgFromNic.virtualNetworkId = suite.isMocked ? output.id : suite.generateId(detachNsgFromNic.virtualNetworkId, null);
-              var cmd = 'network vnet subnet create -g {1} -n subnetName --address-prefix {addressPrefix} --vnet-name virtualNetworkName --json'.formatArgs(subnet, groupName);
+              networkInterfaces.virtualNetworkId = output.id;
+
+              var cmd = 'network vnet subnet create -g {1} -n {name} --address-prefix {addressPrefix} --vnet-name {virtualNetworkName} --json'.formatArgs(subnet, groupName);
               testUtils.executeCommand(suite, retry, cmd, function (result) {
                 result.exitStatus.should.equal(0);
                 var output = JSON.parse(result.text);
-                invalidPrivateIPAddress.subnetId = suite.isMocked ? output.id : suite.generateId(invalidPrivateIPAddress.subnetId, null);
-                ipVersionOutOfRange.subnetId = suite.isMocked ? output.id : suite.generateId(ipVersionOutOfRange.subnetId, null);
-                invalidIPForwarding.subnetId = suite.isMocked ? output.id : suite.generateId(invalidIPForwarding.subnetId, null);
-                attachNsgToNic.subnetId = suite.isMocked ? output.id : suite.generateId(attachNsgToNic.subnetId, null);
-                detachNsgFromNic.subnetId = suite.isMocked ? output.id : suite.generateId(detachNsgFromNic.subnetId, null);
-                var cmd = 'network public-ip create -g {1} -n publicIPAddressName --location {location} --json'.formatArgs(publicIPAddress, groupName);
+                networkInterfaces.subnetId = output.id;
+
+                var cmd = 'network nsg create -g {1} -n {name} --location {location} --json'.formatArgs(networkSecurityGroup, groupName);
                 testUtils.executeCommand(suite, retry, cmd, function (result) {
                   result.exitStatus.should.equal(0);
                   var output = JSON.parse(result.text);
-                  invalidPrivateIPAddress.publicIPAddressId = suite.isMocked ? output.id : suite.generateId(invalidPrivateIPAddress.publicIPAddressId, null);
-                  ipVersionOutOfRange.publicIPAddressId = suite.isMocked ? output.id : suite.generateId(ipVersionOutOfRange.publicIPAddressId, null);
-                  invalidIPForwarding.publicIPAddressId = suite.isMocked ? output.id : suite.generateId(invalidIPForwarding.publicIPAddressId, null);
-                  attachNsgToNic.publicIPAddressId = suite.isMocked ? output.id : suite.generateId(attachNsgToNic.publicIPAddressId, null);
-                  detachNsgFromNic.publicIPAddressId = suite.isMocked ? output.id : suite.generateId(detachNsgFromNic.publicIPAddressId, null);
-                  var cmd = 'network nsg create -g {1} -n networkSecurityGroupName --location {location} --json'.formatArgs(networkSecurityGroup, groupName);
-                  testUtils.executeCommand(suite, retry, cmd, function (result) {
-                    result.exitStatus.should.equal(0);
-                    var output = JSON.parse(result.text);
-                    invalidPrivateIPAddress.networkSecurityGroupId = suite.isMocked ? output.id : suite.generateId(invalidPrivateIPAddress.networkSecurityGroupId, null);
-                    ipVersionOutOfRange.networkSecurityGroupId = suite.isMocked ? output.id : suite.generateId(ipVersionOutOfRange.networkSecurityGroupId, null);
-                    invalidIPForwarding.networkSecurityGroupId = suite.isMocked ? output.id : suite.generateId(invalidIPForwarding.networkSecurityGroupId, null);
-                    attachNsgToNic.networkSecurityGroupId = suite.isMocked ? output.id : suite.generateId(attachNsgToNic.networkSecurityGroupId, null);
-                    detachNsgFromNic.networkSecurityGroupId = suite.isMocked ? output.id : suite.generateId(detachNsgFromNic.networkSecurityGroupId, null);
+                  networkInterfaces.networkSecurityGroupId = output.id;
+
+                  networkTestUtil.createPublicIp(publicIpProp, suite, function (publicIp) {
+                    networkInterfaces.publicIPAddressId = publicIp.id;
                     done();
                   });
                 });
@@ -181,27 +198,17 @@ describe('arm', function () {
           });
         } else {
           var subscriptionId = profile.current.getSubscription().id;
-          invalidPrivateIPAddress.virtualNetworkId = suite.isMocked ? generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'virtualNetworks', invalidPrivateIPAddress.virtualNetworkName) : suite.generateId(invalidPrivateIPAddress.virtualNetworkId, null)
-          invalidPrivateIPAddress.subnetId = suite.isMocked ? generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'subnets', invalidPrivateIPAddress.subnetName) : suite.generateId(invalidPrivateIPAddress.subnetId, null)
-          invalidPrivateIPAddress.publicIPAddressId = suite.isMocked ? generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'publicIPAddresses', invalidPrivateIPAddress.publicIPAddressName) : suite.generateId(invalidPrivateIPAddress.publicIPAddressId, null)
-          ipVersionOutOfRange.virtualNetworkId = suite.isMocked ? generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'virtualNetworks', ipVersionOutOfRange.virtualNetworkName) : suite.generateId(ipVersionOutOfRange.virtualNetworkId, null)
-          ipVersionOutOfRange.subnetId = suite.isMocked ? generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'subnets', ipVersionOutOfRange.subnetName) : suite.generateId(ipVersionOutOfRange.subnetId, null)
-          ipVersionOutOfRange.publicIPAddressId = suite.isMocked ? generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'publicIPAddresses', ipVersionOutOfRange.publicIPAddressName) : suite.generateId(ipVersionOutOfRange.publicIPAddressId, null)
-          invalidIPForwarding.virtualNetworkId = suite.isMocked ? generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'virtualNetworks', invalidIPForwarding.virtualNetworkName) : suite.generateId(invalidIPForwarding.virtualNetworkId, null)
-          invalidIPForwarding.subnetId = suite.isMocked ? generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'subnets', invalidIPForwarding.subnetName) : suite.generateId(invalidIPForwarding.subnetId, null)
-          invalidIPForwarding.publicIPAddressId = suite.isMocked ? generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'publicIPAddresses', invalidIPForwarding.publicIPAddressName) : suite.generateId(invalidIPForwarding.publicIPAddressId, null)
-          attachNsgToNic.virtualNetworkId = suite.isMocked ? generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'virtualNetworks', attachNsgToNic.virtualNetworkName) : suite.generateId(attachNsgToNic.virtualNetworkId, null)
-          attachNsgToNic.subnetId = suite.isMocked ? generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'subnets', attachNsgToNic.subnetName) : suite.generateId(attachNsgToNic.subnetId, null)
-          attachNsgToNic.publicIPAddressId = suite.isMocked ? generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'publicIPAddresses', attachNsgToNic.publicIPAddressName) : suite.generateId(attachNsgToNic.publicIPAddressId, null)
-          attachNsgToNic.networkSecurityGroupId = suite.isMocked ? generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'networkSecurityGroups', attachNsgToNic.networkSecurityGroupName) : suite.generateId(attachNsgToNic.networkSecurityGroupId, null)
-          detachNsgFromNic.networkSecurityGroupId = suite.isMocked ? generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'networkSecurityGroups', detachNsgFromNic.networkSecurityGroupName) : suite.generateId(detachNsgFromNic.networkSecurityGroupId, null)
+          networkInterfaces.virtualNetworkId = generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'virtualNetworks', virtualNetwork.name);
+          networkInterfaces.subnetId = generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'virtualNetworks/' + virtualNetwork.name + '/subnets', subnet.name);
+          networkInterfaces.networkSecurityGroupId = generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'networkSecurityGroups', networkSecurityGroup.name);
+          networkInterfaces.publicIPAddressId = generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'publicIPAddresses', publicIpProp.name);
           done();
         }
       });
     });
     after(function (done) {
-      this.timeout(hour);
-      networkUtil.deleteGroup(groupName, suite, function () {
+      this.timeout(testTimeout);
+      networkTestUtil.deleteGroup(groupName, suite, function () {
         suite.teardownSuite(done);
       });
     });
@@ -212,76 +219,357 @@ describe('arm', function () {
       suite.teardownTest(done);
     });
 
-    describe('network interfaces', function () {
-      this.timeout(hour);
-      it('create should create network interfaces', function (done) {
-        var cmd = 'network nic create -g {group} -n {name} --private-ip-address {privateIPAddress} --private-ip-version {privateIPAddressVersion} --ip-config-name {ipConfigurationName} --internal-dns-name-label {internalDnsNameLabel} --enable-accelerated-networking {enableAcceleratedNetworking} --enable-ip-forwarding {enableIPForwarding} --location {location} --subnet-vnet-name {virtualNetworkName} --subnet-name {subnetName} --public-ip-name {publicIPAddressName} --json'.formatArgs(networkInterfaces);
+    describe('nic custom', function () {
+      this.timeout(testTimeout);
+
+      it('create should create nic with default ip configuration', function (done) {
+        var cmd = ('network nic create -g {group} -n {name} -l {location} -a {privateIPAddress} -r {internalDnsNameLabel} ' +
+          '-f {enableIpForwarding} -t {tags} -u {1} -i {2} --json').formatArgs(networkInterfaces, networkInterfaces.subnetId, networkInterfaces.publicIPAddressId);
         testUtils.executeCommand(suite, retry, cmd, function (result) {
           result.exitStatus.should.equal(0);
           var output = JSON.parse(result.text);
+
           output.name.should.equal(networkInterfaces.name);
-          output.ipConfigurations[index].privateIPAddress.toLowerCase().should.equal(networkInterfaces.privateIPAddress.toLowerCase());
-          output.ipConfigurations[index].privateIPAddressVersion.toLowerCase().should.equal(networkInterfaces.privateIPAddressVersion.toLowerCase());
-          output.ipConfigurations[index].name.toLowerCase().should.equal(networkInterfaces.ipConfigurationName.toLowerCase());
-          output.dnsSettings.internalDnsNameLabel.toLowerCase().should.equal(networkInterfaces.internalDnsNameLabel.toLowerCase());
-          output.enableAcceleratedNetworking.should.equal(utils.parseBool(networkInterfaces.enableAcceleratedNetworking));
-          output.enableIPForwarding.should.equal(utils.parseBool(networkInterfaces.enableIPForwarding));
+          output.enableIPForwarding.should.equal(networkInterfaces.enableIpForwarding);
+          output.ipConfigurations.length.should.equal(1);
+          output.dnsSettings.internalDnsNameLabel.should.equal(networkInterfaces.internalDnsNameLabel);
+
+          var defaultIpConfig = output.ipConfigurations[0];
+          defaultIpConfig.subnet.id.should.containEql(networkInterfaces.subnetName);
+          defaultIpConfig.publicIPAddress.id.should.containEql(publicIpProp.name);
+          defaultIpConfig.privateIPAddress.should.equal(networkInterfaces.privateIPAddress);
+
+          networkTestUtil.shouldBeSucceeded(defaultIpConfig);
+          networkTestUtil.shouldHaveTags(output);
+          networkTestUtil.shouldBeSucceeded(output);
           done();
         });
       });
-      it('show should display network interfaces details', function (done) {
+
+      it('show should display nic details', function (done) {
         var cmd = 'network nic show -g {group} -n {name} --json'.formatArgs(networkInterfaces);
         testUtils.executeCommand(suite, retry, cmd, function (result) {
           result.exitStatus.should.equal(0);
           var output = JSON.parse(result.text);
+
           output.name.should.equal(networkInterfaces.name);
-          output.ipConfigurations[index].privateIPAddress.toLowerCase().should.equal(networkInterfaces.privateIPAddress.toLowerCase());
-          output.ipConfigurations[index].privateIPAddressVersion.toLowerCase().should.equal(networkInterfaces.privateIPAddressVersion.toLowerCase());
-          output.ipConfigurations[index].name.toLowerCase().should.equal(networkInterfaces.ipConfigurationName.toLowerCase());
-          output.dnsSettings.internalDnsNameLabel.toLowerCase().should.equal(networkInterfaces.internalDnsNameLabel.toLowerCase());
-          output.enableAcceleratedNetworking.should.equal(utils.parseBool(networkInterfaces.enableAcceleratedNetworking));
-          output.enableIPForwarding.should.equal(utils.parseBool(networkInterfaces.enableIPForwarding));
+          output.enableIPForwarding.should.equal(networkInterfaces.enableIpForwarding);
+          output.dnsSettings.internalDnsNameLabel.should.equal(networkInterfaces.internalDnsNameLabel);
+          output.ipConfigurations.length.should.equal(1);
+
+          var defaultIpConfig = output.ipConfigurations[0];
+          defaultIpConfig.subnet.id.should.containEql(networkInterfaces.subnetName);
+          defaultIpConfig.publicIPAddress.id.should.containEql(publicIpProp.name);
+          defaultIpConfig.privateIPAddress.should.equal(networkInterfaces.privateIPAddress);
+
+          networkTestUtil.shouldBeSucceeded(defaultIpConfig);
+          networkTestUtil.shouldHaveTags(output);
+          networkTestUtil.shouldBeSucceeded(output);
           done();
         });
       });
-      it('set should update network interfaces', function (done) {
-        var cmd = 'network nic set -g {group} -n {name} --internal-dns-name-label {internalDnsNameLabelNew} --enable-accelerated-networking {enableAcceleratedNetworkingNew} --enable-ip-forwarding {enableIPForwardingNew} --json'.formatArgs(networkInterfaces);
+
+      it('set should modify nic', function (done) {
+        var cmd = ('network nic set -g {group} -n {name} -r {newInternalDnsNameLabel} -f {newEnableIpForwarding} ' +
+          '-t {newTags} --json').formatArgs(networkInterfaces);
         testUtils.executeCommand(suite, retry, cmd, function (result) {
           result.exitStatus.should.equal(0);
           var output = JSON.parse(result.text);
-          output.name.should.equal(networkInterfaces.name);
-          output.dnsSettings.internalDnsNameLabel.toLowerCase().should.equal(networkInterfaces.internalDnsNameLabelNew.toLowerCase());
-          output.enableAcceleratedNetworking.should.equal(utils.parseBool(networkInterfaces.enableAcceleratedNetworkingNew));
-          output.enableIPForwarding.should.equal(utils.parseBool(networkInterfaces.enableIPForwardingNew));
+
+          output.enableIPForwarding.should.equal(networkInterfaces.newEnableIpForwarding);
+          output.dnsSettings.internalDnsNameLabel.should.equal(networkInterfaces.newInternalDnsNameLabel);
+
+          networkTestUtil.shouldAppendTags(output);
+          networkTestUtil.shouldBeSucceeded(output);
           done();
         });
       });
-      it('list should display all network interfaces in resource group', function (done) {
-        var cmd = 'network nic list -g {group} --json'.formatArgs(networkInterfaces);
+
+      it('effective-route-table and effective-nsg commands should work', function (done) {
+        var cmd = 'network nic show -g {group} -n {name} --json'.formatArgs(networkInterfaces);
         testUtils.executeCommand(suite, retry, cmd, function (result) {
           result.exitStatus.should.equal(0);
-          var outputs = JSON.parse(result.text);
-          _.some(outputs, function (output) {
-            return output.name === networkInterfaces.name;
+          var output = JSON.parse(result.text);
+          cmd = 'network nic effective-nsg list {group} {name} --json'.formatArgs(networkInterfaces);
+          testUtils.executeCommand(suite, retry, cmd, function (result) {
+            result.exitStatus.should.not.equal(0);
+            cmd = 'network nic effective-route-table show {group} {name} --json'.formatArgs(networkInterfaces);
+            testUtils.executeCommand(suite, retry, cmd, function (result) {
+              result.exitStatus.should.not.equal(0);
+              vmTestUtil.CreateVmWithNic(networkInterfaces.group, networkInterfaces.attachedVMName, location, 'Linux', 'Debian', output.name, 'xplatuser', 'Pa$$word1', networkInterfaces.attachedVMStorageAccount, suite, function(result) {
+                cmd = 'network nic effective-nsg list {group} {name} --json'.formatArgs(networkInterfaces);
+                testUtils.executeCommand(suite, retry, cmd, function (result) {
+                  result.exitStatus.should.equal(0);
+                  cmd = 'network nic effective-route-table show {group} {name} --json'.formatArgs(networkInterfaces);
+                  testUtils.executeCommand(suite, retry, cmd, function (result) {
+                    result.exitStatus.should.equal(0);
+                    vmTestUtil.RemoveVm(networkInterfaces.group, networkInterfaces.attachedVMName, suite, function(result) {
+                      done();
+                    });
+                  });
+                });
+              });
+            });
+          });
+        });
+      });
+
+      it('ip-config create should should create another ip configuration', function (done) {
+        var cmd = 'network nic ip-config create -g {group} -c {name} -n {ipConfigName} -b {newPrivateIPAddressVersion} --json'.formatArgs(networkInterfaces);
+
+        testUtils.executeCommand(suite, retry, cmd, function (result) {
+          result.exitStatus.should.equal(0);
+          var output = JSON.parse(result.text);
+
+          output.ipConfigurations.length.should.equal(2);
+
+          var anotherIpConfig = output.ipConfigurations[1];
+          anotherIpConfig.name.should.equal(networkInterfaces.ipConfigName);
+          anotherIpConfig.privateIPAddressVersion.should.equal(networkInterfaces.newPrivateIPAddressVersion);
+
+          networkTestUtil.shouldBeSucceeded(anotherIpConfig);
+          networkTestUtil.shouldBeSucceeded(output);
+          done();
+        });
+      });
+
+      it('ip-config list should should list ip configurations in nic', function (done) {
+        var cmd = 'network nic ip-config list -g {group} -c {name} --json'.formatArgs(networkInterfaces);
+
+        testUtils.executeCommand(suite, retry, cmd, function (result) {
+          result.exitStatus.should.equal(0);
+          var output = JSON.parse(result.text);
+
+          output.length.should.equal(2);
+          _.some(output, function (ipConfig) {
+            return ipConfig.name === networkInterfaces.ipConfigName;
           }).should.be.true;
+
           done();
         });
       });
-      it('effective-route-table should fail to perform get effective route table operation successfully', function (done) {
-        var cmd = 'network nic effective-route-table show -g {group} --nic-name {networkInterfaceName} --json'.formatArgs(networkInterfaces);
+
+      it('ip-config show should display details of ip configuration', function (done) {
+        var cmd = 'network nic ip-config show -g {group} -c {name} -n {ipConfigName} --json'.formatArgs(networkInterfaces);
+
         testUtils.executeCommand(suite, retry, cmd, function (result) {
-          result.exitStatus.should.not.equal(0);
+          result.exitStatus.should.equal(0);
+          var output = JSON.parse(result.text);
+
+          output.name.should.equal(networkInterfaces.ipConfigName);
+          output.privateIPAddressVersion.should.equal(networkInterfaces.newPrivateIPAddressVersion);
+
+          networkTestUtil.shouldBeSucceeded(output);
           done();
         });
       });
-      it('effective-nsg should fail to perform list effective network security groups operation successfully', function (done) {
-        var cmd = 'network nic effective-nsg list -g {group} --nic-name {networkInterfaceName} --json'.formatArgs(networkInterfaces);
+
+      it('ip-config delete should delete ip configuration', function (done) {
+        var cmd = 'network nic ip-config delete -g {group} -c {name} -n {ipConfigName} --quiet --json'.formatArgs(networkInterfaces);
         testUtils.executeCommand(suite, retry, cmd, function (result) {
-          result.exitStatus.should.not.equal(0);
+          result.exitStatus.should.equal(0);
+
+          cmd = 'network nic ip-config show -g {group} -c {name} -n {ipConfigName} --json'.formatArgs(networkInterfaces);
+          testUtils.executeCommand(suite, retry, cmd, function (result) {
+            result.exitStatus.should.equal(0);
+            var output = JSON.parse(result.text);
+            output.should.be.empty;
+            done();
+          });
+        });
+      });
+
+      it('ip-config set should attach address pool to ip configuration', function (done) {
+        networkTestUtil.createLB(lbProp, suite, function () {
+          networkTestUtil.createAddressPool(groupName, lbProp.name, poolName, suite, function (pool) {
+            var cmd = 'network nic ip-config set -g {group} -c {name} -n default-ip-config -d {1} --json'.formatArgs(networkInterfaces, pool.id);
+            testUtils.executeCommand(suite, retry, cmd, function (result) {
+              result.exitStatus.should.equal(0);
+              var output = JSON.parse(result.text);
+
+              var defaultIpConfig = output.ipConfigurations[0];
+              var pools = defaultIpConfig.loadBalancerBackendAddressPools;
+              pools.length.should.equal(1);
+              pools[0].id.should.equal(pool.id);
+
+              done();
+            });
+          });
+        });
+      });
+
+      it('ip-config set should detach address pool from ip configuration', function (done) {
+        var cmd = 'network nic ip-config set -g {group} -c {name} -n default-ip-config -d --json'.formatArgs(networkInterfaces);
+        testUtils.executeCommand(suite, retry, cmd, function (result) {
+          result.exitStatus.should.equal(0);
+          var output = JSON.parse(result.text);
+
+          output.ipConfigurations.length.should.equal(1);
+          output.ipConfigurations[0].should.not.have.property('loadBalancerBackendAddressPools');
+
           done();
         });
       });
-      it('delete should delete network interfaces', function (done) {
+
+      it('ip-config set should attach inbound nat rule to ip configuration', function (done) {
+        networkTestUtil.createInboundNatRule(groupName, lbProp.name, inboundRuleName, protocol, frontendPort, backendPort, 'true', 10, lbProp.fipName, suite, function (natRule) {
+          var cmd = 'network nic ip-config set -g {group} -c {name} -n default-ip-config -e {1} --json'.formatArgs(networkInterfaces, natRule.id);
+          testUtils.executeCommand(suite, retry, cmd, function (result) {
+            result.exitStatus.should.equal(0);
+            var output = JSON.parse(result.text);
+
+            var rules = output.ipConfigurations[0].loadBalancerInboundNatRules;
+            rules.length.should.equal(1);
+            rules[0].id.should.equal(natRule.id);
+
+            done();
+          });
+        });
+      });
+
+      it('ip-config set should detach inbound nat rule from ip configuration', function (done) {
+        var cmd = 'network nic ip-config set -g {group} -c {name} -n default-ip-config -e --json'.formatArgs(networkInterfaces);
+        testUtils.executeCommand(suite, retry, cmd, function (result) {
+          result.exitStatus.should.equal(0);
+          var output = JSON.parse(result.text);
+
+          output.ipConfigurations.length.should.equal(1);
+          output.ipConfigurations[0].should.not.have.property('loadBalancerInboundNatRules');
+
+          done();
+        });
+      });
+
+      it('ip-config address-pool create should attach address pool to ip configuration', function (done) {
+        var cmd = 'network nic ip-config address-pool create -g {group} -c {name} -n default-ip-config -l {1} -a {2} --json'.formatArgs(networkInterfaces, lbProp.name, poolName);
+        testUtils.executeCommand(suite, retry, cmd, function (result) {
+          result.exitStatus.should.equal(0);
+          var output = JSON.parse(result.text);
+
+          output.ipConfigurations.length.should.equal(1);
+          var pools = output.ipConfigurations[0].loadBalancerBackendAddressPools;
+          pools.length.should.equal(1);
+          pools[0].id.should.containEql(poolName);
+
+          done();
+        });
+      });
+
+      it('ip-config address-pool delete should detach address-pool from ip configuration', function (done) {
+        var cmd = 'network nic ip-config address-pool delete -g {group} -c {name} -n default-ip-config -l {1} -a {2} --json'.formatArgs(networkInterfaces, lbProp.name, poolName);
+        testUtils.executeCommand(suite, retry, cmd, function (result) {
+          result.exitStatus.should.equal(0);
+          var output = JSON.parse(result.text);
+
+          output.ipConfigurations.length.should.equal(1);
+          output.ipConfigurations[0].should.not.have.property('loadBalancerBackendAddressPools');
+
+          done();
+        });
+      });
+
+      it('ip-config inbound-nat-rule create should attach inbound nat rule to ip configuration', function (done) {
+        var cmd = 'network nic ip-config inbound-nat-rule create -g {group} -c {name} -n default-ip-config -l {1} -r {2} --json'.formatArgs(networkInterfaces, lbProp.name, inboundRuleName);
+        testUtils.executeCommand(suite, retry, cmd, function (result) {
+          result.exitStatus.should.equal(0);
+          var output = JSON.parse(result.text);
+
+          output.ipConfigurations.length.should.equal(1);
+          var rules = output.ipConfigurations[0].loadBalancerInboundNatRules;
+          rules.length.should.equal(1);
+          rules[0].id.should.containEql(inboundRuleName);
+
+          done();
+        });
+      });
+
+      it('ip-config inbound-nat-rule delete should detach inbound nat rule from ip configuration', function (done) {
+        var cmd = 'network nic ip-config inbound-nat-rule delete -g {group} -c {name} -n default-ip-config -l {1} -r {2} --json'.formatArgs(networkInterfaces, lbProp.name, inboundRuleName);
+        testUtils.executeCommand(suite, retry, cmd, function (result) {
+          result.exitStatus.should.equal(0);
+          var output = JSON.parse(result.text);
+
+          output.ipConfigurations.length.should.equal(1);
+          output.ipConfigurations[0].should.not.have.property('loadBalancerInboundNatRules');
+
+          done();
+        });
+      });
+
+      it('ip-config set should detach public ip from ip configuration', function (done) {
+        var cmd = 'network nic ip-config set -g {group} -c {name} -n default-ip-config --public-ip-name \'\' --json'.formatArgs(networkInterfaces);
+        testUtils.executeCommand(suite, retry, cmd, function (result) {
+          result.exitStatus.should.equal(0);
+          var output = JSON.parse(result.text);
+
+          output.ipConfigurations.length.should.equal(1);
+          output.ipConfigurations[0].should.not.have.property('publicIPAddress');
+
+          done();
+        });
+      });
+
+      it('ip-config create should create second ip configuration', function (done) {
+        var cmd = 'network nic ip-config create -g {group} -c {nicName} -n {name} --json'.formatArgs(ipConfigProp1);
+        testUtils.executeCommand(suite, retry, cmd, function (result) {
+          result.exitStatus.should.equal(0);
+          var output = JSON.parse(result.text);
+
+          var createdConfig = utils.findFirstCaseIgnore(output.ipConfigurations, {name: ipConfigProp1.name});
+          createdConfig.should.not.be.empty;
+
+          networkTestUtil.shouldBeSucceeded(createdConfig);
+          done();
+        });
+      });
+
+      it('ip-config create should create third ip configuration', function (done) {
+        var cmd = 'network nic ip-config create -g {group} -c {nicName} -n {name} -m {vnetName} -k {subnetName} -y {isPrimary} -q --json'.formatArgs(ipConfigProp2);
+        testUtils.executeCommand(suite, retry, cmd, function (result) {
+          result.exitStatus.should.equal(0);
+          var output = JSON.parse(result.text);
+
+          var createdConfig = utils.findFirstCaseIgnore(output.ipConfigurations, {name: ipConfigProp2.name});
+          createdConfig.should.not.be.empty;
+          createdConfig.primary.should.be.true;
+
+          _.some(output.ipConfigurations, function (ipConfig) {
+            return ipConfig.isPrimary === true && ipConfig.name === ipConfigProp2.name;
+          }).should.be.false;
+
+          networkTestUtil.shouldBeSucceeded(createdConfig);
+          done();
+        });
+      });
+
+      it('ip-config set should set second ip configuration as primary', function (done) {
+        var cmd = 'network nic ip-config set -g {group} -c {nicName} -n {name} -y true --json'.formatArgs(ipConfigProp1);
+        testUtils.executeCommand(suite, retry, cmd, function (result) {
+          result.exitStatus.should.equal(0);
+          var output = JSON.parse(result.text);
+
+          var primaryConfig = utils.findFirstCaseIgnore(output.ipConfigurations, {name: ipConfigProp1.name});
+          primaryConfig.primary.should.be.true;
+
+          done();
+        });
+      });
+
+      it('ip-config delete should delete third ip configuration', function (done) {
+        var cmd = 'network nic ip-config delete -g {group} -c {nicName} -n {name} -q --json'.formatArgs(ipConfigProp2);
+        testUtils.executeCommand(suite, retry, cmd, function (result) {
+          result.exitStatus.should.equal(0);
+          var output = JSON.parse(result.text);
+
+          _.some(output.ipConfigurations, function (ipConfig) {
+            return ipConfig.name === ipConfigProp2.name;
+          }).should.be.false;
+          
+          done();
+        });
+      });
+
+      it('delete should delete nic', function (done) {
         var cmd = 'network nic delete -g {group} -n {name} --quiet --json'.formatArgs(networkInterfaces);
         testUtils.executeCommand(suite, retry, cmd, function (result) {
           result.exitStatus.should.equal(0);
@@ -289,65 +577,44 @@ describe('arm', function () {
           cmd = 'network nic show -g {group} -n {name} --json'.formatArgs(networkInterfaces);
           testUtils.executeCommand(suite, retry, cmd, function (result) {
             result.exitStatus.should.equal(0);
-            var output = JSON.parse(result.text || '{}');
+            var output = JSON.parse(result.text);
             output.should.be.empty;
             done();
           });
         });
       });
-      it('create should fail for invalid private ip address', function (done) {
-        var cmd = 'network nic create -g {group} -n {name} --private-ip-address {privateIPAddress} --location {location} --subnet-vnet-name {virtualNetworkName} --subnet-name {subnetName} --public-ip-name {publicIPAddressName} --json'.formatArgs(invalidPrivateIPAddress);
-        testUtils.executeCommand(suite, retry, cmd, function (result) {
-          result.exitStatus.should.not.equal(0);
+
+      it('create vmss should create virtual machine scale set', function (done) {
+        vmTestUtil.createVMSSWithParamFile(networkInterfaces, vmTestUtilCreateVMSSWithParamFile, groupName, suite, function (result) {
+          result.exitStatus.should.equal(0);
           done();
         });
       });
-      it('create should fail for ip version out of range', function (done) {
-        var cmd = 'network nic create -g {group} -n {name} --private-ip-version {privateIPAddressVersion} --location {location} --subnet-vnet-name {virtualNetworkName} --subnet-name {subnetName} --public-ip-name {publicIPAddressName} --json'.formatArgs(ipVersionOutOfRange);
-        testUtils.executeCommand(suite, retry, cmd, function (result) {
-          result.exitStatus.should.not.equal(0);
-          done();
-        });
-      });
-      it('create should fail for invalid ip forwarding', function (done) {
-        var cmd = 'network nic create -g {group} -n {name} --enable-ip-forwarding {enableIPForwarding} --location {location} --subnet-vnet-name {virtualNetworkName} --subnet-name {subnetName} --public-ip-name {publicIPAddressName} --json'.formatArgs(invalidIPForwarding);
-        testUtils.executeCommand(suite, retry, cmd, function (result) {
-          result.exitStatus.should.not.equal(0);
-          done();
-        });
-      });
-      it('attach nsg to nic should pass', function (done) {
-        var cmd = 'network nic create -g {group} -n {name} --location {location} --subnet-vnet-name {virtualNetworkName} --subnet-name {subnetName} --public-ip-name {publicIPAddressName} --json'.formatArgs(attachNsgToNic);
+
+      it('show vmss should display network interface in vmss', function (done) {
+        var cmd = ('network nic show -g {group} -n {networkInterfaceName} ' +
+          '--virtual-machine-scale-set-name {vmssName} ' +
+          '--virtual-machine-index {virtualmachineIndex} --json').formatArgs(networkInterfaces);
         testUtils.executeCommand(suite, retry, cmd, function (result) {
           result.exitStatus.should.equal(0);
           var output = JSON.parse(result.text);
-          output.name.should.equal(attachNsgToNic.name);
-          output.ipConfigurations[index].subnet.id.should.containEql(attachNsgToNic.subnetName);
-          output.ipConfigurations[index].publicIPAddress.id.should.containEql(attachNsgToNic.publicIPAddressName);
-
-          cmd = 'network nic set -g {group} -n {name} --network-security-group-name {networkSecurityGroupName} --json'.formatArgs(attachNsgToNic);
-          testUtils.executeCommand(suite, retry, cmd, function (result) {
-            result.exitStatus.should.equal(0);
-            var output = JSON.parse(result.text);
-            output.name.should.equal(attachNsgToNic.name);
-            output.networkSecurityGroup.id.should.containEql(attachNsgToNic.networkSecurityGroupName);
-            done();
-          });
+          output.name.should.equal(networkInterfaces.networkInterfaceName);
+          done();
         });
       });
-      it('detach nsg from nic should pass', function (done) {
-        var cmd = 'network nic set -g {group} -n {name} --network-security-group-name --json'.formatArgs(detachNsgFromNic);
+
+      it('list vmss should list all network interfaces in vmss', function (done) {
+        var cmd = ('network nic list -g {group} --virtual-machine-scale-set-name {vmssName} ' +
+          '--virtual-machine-index {virtualmachineIndex} --json').formatArgs(networkInterfaces);
         testUtils.executeCommand(suite, retry, cmd, function (result) {
           result.exitStatus.should.equal(0);
           var output = JSON.parse(result.text);
-          output.name.should.equal(detachNsgFromNic.name);
-          should.not.exist(output.networkSecurityGroup);
 
-          cmd = 'network nic delete -g {group} -n {name} --quiet --json'.formatArgs(detachNsgFromNic);
-          testUtils.executeCommand(suite, retry, cmd, function (result) {
-            result.exitStatus.should.equal(0);
-            done();
-          });
+          _.some(output, function (nic) {
+            return nic.name === networkInterfaces.networkInterfaceName;
+          }).should.be.true;
+
+          done();
         });
       });
     });
