+++ test/recordings/arm-network-nic-ip-config-tests-generated/arm_network_ip_configurations_create_should_create_ip_configurations.nock.js
