+++ test/recordings/arm-network-nic-ip-config-tests-generated/arm_network_ip_configurations_delete_should_delete_ip_configurations.nock.js
