+++ test/recordings/arm-network-nic-ip-config-tests-generated/arm_network_ip_configurations_list_should_display_all_ip_configurations_in_resource_group.nock.js
