+++ test/recordings/arm-network-nic-ip-config-tests-generated/arm_network_ip_configurations_show_should_display_ip_configurations_details.nock.js
