+++ test/recordings/arm-network-nic-ip-config-tests-generated/suite.arm-network-nic-ip-config-tests-generated.nock.js
