+++ test/recordings/arm-network-nic-tests-generated/arm_network_network_interfaces_attach_nsg_to_nic_should_pass.nock.js
