+++ test/recordings/arm-network-nic-tests-generated/arm_network_network_interfaces_create_should_create_network_interfaces.nock.js
