+++ test/recordings/arm-network-nic-tests-generated/arm_network_network_interfaces_create_should_fail_for_invalid_private_ip_address.nock.js
