+++ test/recordings/arm-network-nic-tests-generated/arm_network_network_interfaces_delete_should_delete_network_interfaces.nock.js
