+++ test/recordings/arm-network-nic-tests-generated/arm_network_network_interfaces_detach_nsg_from_nic_should_pass.nock.js
