+++ test/recordings/arm-network-nic-tests-generated/arm_network_network_interfaces_effective-nsg_should_fail_to_perform_list_effective_network_security_groups_operation_successfully.nock.js
