+++ test/recordings/arm-network-nic-tests-generated/arm_network_network_interfaces_effective-route-table_should_fail_to_perform_get_effective_route_table_operation_successfully.nock.js
