+++ test/recordings/arm-network-nic-tests-generated/arm_network_network_interfaces_list_should_display_all_network_interfaces_in_resource_group.nock.js
