+++ test/recordings/arm-network-nic-tests-generated/arm_network_network_interfaces_set_should_update_network_interfaces.nock.js
