+++ test/recordings/arm-network-nic-tests-generated/arm_network_network_interfaces_show_should_display_network_interfaces_details.nock.js
