+++ test/recordings/arm-network-nic-tests-generated/arm_network_nic_effective-route-table_and_effective-nsg_commands_should_work.nock.js
