+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_create_should_create_nic_with_default_ip_configuration.nock.js
