+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_create_vmss_should_create_virtual_machine_scale_set.nock.js
