+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_delete_should_delete_nic.nock.js
