+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_effective-route-table_and_effective-nsg_commands_should_work.nock.js
