+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_ip-config_address-pool_delete_should_detach_address-pool_from_ip_configuration.nock.js
