+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_ip-config_create_should_create_third_ip_configuration.nock.js
