+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_ip-config_delete_should_delete_third_ip_configuration.nock.js
