+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_ip-config_list_should_should_list_ip_configurations_in_nic.nock.js
