+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_ip-config_set_should_attach_inbound_nat_rule_to_ip_configuration.nock.js
