+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_ip-config_set_should_detach_address_pool_from_ip_configuration.nock.js
