+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_ip-config_set_should_set_second_ip_configuration_as_primary.nock.js
