+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_ip-config_show_should_display_details_of_ip_configuration.nock.js
