+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_list_vmss_should_list_all_network_interfaces_in_vmss.nock.js
