+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_set_should_modify_nic.nock.js
