+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_show_should_display_nic_details.nock.js
