+++ test/recordings/arm-network-nic-tests/arm_network_nic_custom_show_vmss_should_display_network_interface_in_vmss.nock.js
